\name{actData}
\alias{actData}
\docType{data}
\title{
Sample Action Data
}
\description{
Sample Action Data. x1 ~ x17 is the explanatory variable, y is the target variable.
}

\usage{data(actData)}
\format{
  A data frame with 500 observations on the following 18 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{x5}}{a numeric vector}
    \item{\code{x6}}{a numeric vector}
    \item{\code{x7}}{a numeric vector}
    \item{\code{x8}}{a numeric vector}
    \item{\code{x9}}{a numeric vector}
    \item{\code{x10}}{a numeric vector}
    \item{\code{x11}}{a numeric vector}
    \item{\code{x12}}{a numeric vector}
    \item{\code{x13}}{a numeric vector}
    \item{\code{x14}}{a numeric vector}
    \item{\code{x15}}{a numeric vector}
    \item{\code{x16}}{a numeric vector}
    \item{\code{x17}}{a numeric vector}
  }
}


\examples{
data(actData)
## maybe str(actData) ; plot(actData) ...
}
\keyword{datasets}
